+++ test/coda/on_this_day_test.exs
@@ -1,8 +1,6 @@
 defmodule Coda.OnThisDayTest do
   use ExUnit.Case, async: true
-
-  import Fixtures.Archive
-  import Fixtures.Lastfm
+  import Coda.Factory
 
   alias Explorer.DataFrame
   alias Coda.OnThisDay
@@ -10,13 +8,16 @@ defmodule Coda.OnThisDayTest do
   setup_all do
     %{
       dataframe:
-        "a_user" |> recent_tracks_on_this_day() |> dataframe() |> DataFrame.rename(name: "track")
+        build(:scrobbles, rows: 10)
+        |> Enum.map(&Map.from_struct/1)
+        |> DataFrame.new(lazy: true)
+        |> DataFrame.rename(name: "track")
     }
   end
 
   test "render_overview/1", %{dataframe: df} do
     assert %Kino.Markdown{content: content} = OnThisDay.render_overview(df)
-    assert content =~ "**1** scrobbles"
+    assert content =~ "**10** scrobbles"
   end
 
   test "render_most_played/1", %{dataframe: df} do