+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0 (2023-09-13)
+
+* faster (optimised) on-this-day data frame creation functions and Livebook UX
+* refactor `Analytics`, `Livebook` behaviours and default implementation
+* refactor Livebook rendering functions
+* introduce `ExMachina` test data factory with 200 test track samples
+
 ## 0.1.0 (2023-09-01)
 
 * initial commits: on-this-day and common analytics features migrated from [lastfm_archive 0.11.1](https://github.com/boonious/lastfm_archive/blob/master/CHANGELOG.md#0111-2023-08-21)